Set up a store
Many errors are present.

Due to typescript unable to find dev tools I created a gloabl types declation file and even added it to tsconfig file, but it is still not being recognized by typescript server. Current solution has been to redeclare in store file.

There are also some type errors present due to the opposite right-to-left flow of the compose function.